ml-model/insights-api/tests/test_database.py
import pytest
from database import _get_conn

def test_db_connection_fail_gracefully():
    # Without a valid DATABASE_URL in the env, this should raise an exception or return None seamlessly
    # The application is designed to log a warning and retry later
    try:
        conn = _get_conn()
        if conn:
            conn.close()
    except Exception as e:
        # Expected if DATABASE_URL is missing or invalid
        pass
