ml-model/insights-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
earthengine-api>=0.1.390
numpy>=1.24.0
scipy>=1.11.0
shapely>=2.0.0
geojson>=3.1.0
psycopg2-binary>=2.9.9
cloudinary>=1.36.0
groq>=0.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
Pillow>=10.0.0
